requirements.txt
datasets
evaluate
fastapi
numpy
openai_whisper
pyarrow
SQLAlchemy
torch
torchaudio
tqdm
transformers
